+++ src/Symfony/Component/Form/Tests/CompoundFormTest.php
@@ -33,7 +33,7 @@ class CompoundFormTest extends AbstractFormTest
         $this->assertTrue($this->form->isValid());
     }
 
-    public function testInvalidIfChildrenIsInvalid()
+    public function testInvalidIfChildIsInvalid()
     {
         $this->form->add($this->getValidForm('firstName'));
         $this->form->add($this->getInvalidForm('lastName'));
@@ -135,12 +135,71 @@ class CompoundFormTest extends AbstractFormTest
         $child = $this->getBuilder('foo')->getForm();
         $this->form->add($child);
 
+        $this->assertTrue($this->form->has('foo'));
+        $this->assertSame($this->form, $child->getParent());
+        $this->assertSame(array('foo' => $child), $this->form->all());
+    }
+
+    public function testAddUsingNameAndType()
+    {
+        $child = $this->getBuilder('foo')->getForm();
+
+        $this->factory->expects($this->once())
+            ->method('createNamed')
+            ->with('foo', 'text', null, array('bar' => 'baz'))
+            ->will($this->returnValue($child));
+
+        $this->form->add('foo', 'text', array('bar' => 'baz'));
+
+        $this->assertTrue($this->form->has('foo'));
+        $this->assertSame($this->form, $child->getParent());
+        $this->assertSame(array('foo' => $child), $this->form->all());
+    }
+
+    public function testAddUsingNameButNoType()
+    {
+        $this->form = $this->getBuilder('name', null, '\stdClass')
+            ->setCompound(true)
+            ->setDataMapper($this->getDataMapper())
+            ->getForm();
+
+        $child = $this->getBuilder('foo')->getForm();
+
+        $this->factory->expects($this->once())
+            ->method('createForProperty')
+            ->with('\stdClass', 'foo')
+            ->will($this->returnValue($child));
+
+        $this->form->add('foo');
+
+        $this->assertTrue($this->form->has('foo'));
+        $this->assertSame($this->form, $child->getParent());
+        $this->assertSame(array('foo' => $child), $this->form->all());
+    }
+
+    public function testAddUsingNameButNoTypeAndOptions()
+    {
+        $this->form = $this->getBuilder('name', null, '\stdClass')
+            ->setCompound(true)
+            ->setDataMapper($this->getDataMapper())
+            ->getForm();
+
+        $child = $this->getBuilder('foo')->getForm();
+
+        $this->factory->expects($this->once())
+            ->method('createForProperty')
+            ->with('\stdClass', 'foo', null, array('bar' => 'baz'))
+            ->will($this->returnValue($child));
+
+        $this->form->add('foo', null, array('bar' => 'baz'));
+
+        $this->assertTrue($this->form->has('foo'));
         $this->assertSame($this->form, $child->getParent());
         $this->assertSame(array('foo' => $child), $this->form->all());
     }
 
     /**
-     * @expectedException Symfony\Component\Form\Exception\AlreadyBoundException
+     * @expectedException \Symfony\Component\Form\Exception\AlreadyBoundException
      */
     public function testAddThrowsExceptionIfAlreadyBound()
     {
@@ -161,7 +220,7 @@ class CompoundFormTest extends AbstractFormTest
     }
 
     /**
-     * @expectedException Symfony\Component\Form\Exception\AlreadyBoundException
+     * @expectedException \Symfony\Component\Form\Exception\AlreadyBoundException
      */
     public function testRemoveThrowsExceptionIfAlreadyBound()
     {